+++ src/main/java/Springweb/controller/VegetableController.java
@@ -1,9 +1,14 @@
 
 package Springweb.controller;
 
+import Springweb.entity.Category;
+import Springweb.entity.Orderdetail;
 import Springweb.entity.Vegetable;
+import Springweb.repository.CategoryRepository;
+import Springweb.repository.OrderDetailRepository;
 
 import Springweb.repository.VegetableRepository;
+import java.util.ArrayList;
 import java.util.Optional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
@@ -18,6 +23,13 @@ public class VegetableController {
 
     @Autowired
     VegetableRepository vegetableRepository;
+    
+    @Autowired
+    private  OrderDetailRepository orderDetailRepository;
+    
+    @Autowired
+    private CategoryRepository categoryRepository;
+    
     @GetMapping("admin/vegetables")
     public String showAllCategories(Model m) {
         Iterable<Vegetable> list = vegetableRepository.findAll();
@@ -29,7 +41,10 @@ public class VegetableController {
     @GetMapping("admin/vegetables/add")
     public String create(Model m) {
         Vegetable veg = new Vegetable();
-        m.addAttribute("vegetable", veg);
+        ArrayList<Category> list = (ArrayList<Category>) categoryRepository.findAll();
+        m.addAttribute("vegetable", veg);   
+        m.addAttribute("category", list);
+
         return "admin/vegetables/vegetable_add";
 
     }
@@ -52,7 +67,11 @@ public class VegetableController {
         c.setVegetable_Name(vegetable.getVegetable_Name());
         c.setUnit(vegetable.getUnit());
         c.setAmount(vegetable.getAmount());
-        c.setImage(vegetable.getImage());
+        if(vegetable.getImage().isEmpty()) {
+            c.setImage(c.getImage());
+        }else {
+            c.setImage(vegetable.getImage());
+        }
         c.setPrice(vegetable.getPrice());
         vegetableRepository.save(c);
 
@@ -63,11 +82,18 @@ public class VegetableController {
     @GetMapping(value = {"admin/vegetables/edit/{id}"})
     public String edit(@PathVariable("id") int id, Model model) {
         Optional<Vegetable> veg = vegetableRepository.findById(id);
+        ArrayList<Category> list = (ArrayList<Category>) categoryRepository.findAll();
+        model.addAttribute("category", list);
         veg.ifPresent(vegetable -> model.addAttribute("vegetable", veg));
         return "admin/vegetables/vegetable_edit";
     }
     @GetMapping(value = {"admin/vegetables/delete/{id}"})
     public String delete(@PathVariable("id") int id, Model model){
+        ArrayList<Orderdetail> list = (ArrayList<Orderdetail>) orderDetailRepository.CheckForeignKey(id);
+        if(!list.isEmpty()) {
+            model.addAttribute("flag", "Exit in orther table");
+            return "redirect:/admin/vegetables";
+        }
         vegetableRepository.deleteById(id);
         model.addAttribute("vegetable", vegetableRepository.findAll());
         return "redirect:/admin/vegetables";
